Organize “item”-related files in a folder per ng style guide
