+++ app.module.ts
@@ -3,9 +3,9 @@ import { NativeScriptModule } from "nativescript-angular/nativescript.module";
 import { AppRoutingModule } from "./app.routing";
 import { AppComponent } from "./app.component";
 
-import { ItemService } from "./item.service";
-import { ItemsComponent } from "./items.component";
-import { ItemDetailComponent } from "./item-detail.component";
+import { ItemService } from "./item/item.service";
+import { ItemsComponent } from "./item/items.component";
+import { ItemDetailComponent } from "./item/item-detail.component";
 
 @NgModule({
     bootstrap: [
