+++ app.routing.ts
@@ -2,8 +2,8 @@ import { NgModule } from "@angular/core";
 import { NativeScriptRouterModule } from "nativescript-angular/router";
 import { Routes } from "@angular/router";
 
-import { ItemsComponent } from "./items.component";
-import { ItemDetailComponent } from "./item-detail.component";
+import { ItemsComponent } from "./item/items.component";
+import { ItemDetailComponent } from "./item/item-detail.component";
 
 const routes: Routes = [
     { path: "", redirectTo: "/items", pathMatch: "full" },
