--- item.ts
@@ -1,5 +1,0 @@
-export class Item {
-  id: number;
-  name: string;
-  role: string;
-}
