+++ item/item.ts
@@ -1,0 +1,5 @@
+export class Item {
+    id: number;
+    name: string;
+    role: string;
+}
